Fix VN-200 checksum bug on amd64.
Fixes a bug in the new VN-200 CRC16 checksum implementation on (at
least) the amd64 architecture. With the previous implementation, testing
found that the checksum was always failing for valid packets because of
an integer size issue when obtaining the i-th byte of the QByteArray,
which was being converted to an int. This size issue was resolved by
static casting the byte to a quint8. This did not manifest in the old
VN-200 parsing code as the packet was stored as an array of quint8s.

Also changes the casts in the validateVN200Checksum function to C++
style static_casts instead of C-style casts.

+++ src/vn200/vn200.cc
@@ -128,9 +128,9 @@ validateVN200Checksum(QByteArray pkt)
     quint16 crc = 0;
     // Calculate checksum.
     for (quint8 i = 1; i < pkt.size(); ++i) {
-        crc = (quint8) (crc >> 8) | (crc << 8);
-        crc ^= pkt[i];
-        crc ^= (quint8) (crc & 0xff) >> 4;
+        crc = static_cast<quint8>(crc >> 8) | (crc << 8);
+        crc ^= static_cast<quint8>(pkt[i]);
+        crc ^= static_cast<quint8>(crc & 0xff) >> 4;
         crc ^= crc << 12;
         crc ^= (crc & 0x00ff) << 5;
     }
